Add authorization to events, users, invites, and comments

+++ api/lib/api_web/controllers/comment_controller.ex
@@ -9,6 +9,21 @@ defmodule ApiWeb.CommentController do
 
   plug Plugs.RequireAuth, "en" when action in [:index, :create, :show, :update, :delete]
 
+  # Checks if the authenticated user is the owner of the given comment
+  def comment_owner?(conn, comment) do
+    logged_in_user_id = conn.assigns[:user].id
+    owner_user_id = comment.user_id
+    logged_in_user_id == owner_user_id
+  end
+
+  # Checks if the authenticated user is the owner of the given comment or event
+  def comment_or_event_owner?(conn, comment) do
+    logged_in_user_id = conn.assigns[:user].id
+
+    comment_owner?(conn, comment) || 
+    comment.event.user_id == logged_in_user_id
+  end
+
   def index(conn, _params) do
     comments = Comments.list_comments()
     render(conn, "index.json", comments: comments)
@@ -45,8 +60,23 @@ defmodule ApiWeb.CommentController do
   def delete(conn, %{"id" => id}) do
     comment = Comments.get_comment!(id)
 
-    with {:ok, %Comment{}} <- Comments.delete_comment(comment) do
-      send_resp(conn, :no_content, "")
+    # Ensure that the user is the owner of the event or the comment
+    if comment_or_event_owner?(conn, comment) do
+      with {:ok, %Comment{}} <- Comments.delete_comment(comment) do
+        send_resp(conn, :no_content, "")
+      end
+    else 
+      conn
+      |> put_resp_header(
+        "content-type",
+        "application/json; charset=UTF-8")
+      |> send_resp(
+        :unauthorized,
+        Jason.encode!(%{error: "Sorry, you need to be the host or owner of the comment."})
+      )
     end
+
+  
   end
+
 end
